+++ packages/types/src/documents/app/index.ts
@@ -1,5 +1,4 @@
 export * from "../document"
-export * from "./app"
 export * from "./automation"
 export * from "./backup"
 export * from "./component"
@@ -9,7 +8,6 @@ export * from "./layout"
 export * from "./links"
 export * from "./metadata"
 export * from "./oauth2"
-export * from "./workspaceApp"
 export * from "./query"
 export * from "./role"
 export * from "./row"
@@ -22,4 +20,6 @@ export * from "./theme"
 export * from "./user"
 export * from "./view"
 export * from "./webhook"
+export * from "./workspace"
+export * from "./workspaceApp"
 export * from "./workspaceFavourite"
